separated accept offer/counter offer from offer.js
