+++ src/main/resources/static/css/output/outputcss.css
@@ -698,11 +698,6 @@ Ensure the default browser behavior of the `hidden` attribute.
   border-color: rgb(128 128 216 / var(--tw-border-opacity)) !important;
 }
 
-.border-black {
-  --tw-border-opacity: 1 !important;
-  border-color: rgb(0 0 0 / var(--tw-border-opacity)) !important;
-}
-
 .bg-secondary {
   --tw-bg-opacity: 1 !important;
   background-color: rgb(154 177 194 / var(--tw-bg-opacity)) !important;
