+++ src/main/resources/static/js/views/Offers.js
@@ -1,4 +1,7 @@
 import createView from "../createView.js";
+import {initCounterOffer} from "./counterOffer.js";
+import {updateListingObject, updateOfferStatus, confirmOfferAcceptance} from "./acceptOffer.js";
+
 
 const OFFERS_URL = `http://${BACKEND_HOST}:${PORT}/api/offers`;
 
@@ -42,6 +45,7 @@ const retrieveOffersFromDb = (offers) => {
     idArray = [];
     offers.map(offer => idArray.push(offer.id));
     console.log(idArray);
+
     // language=HTML
     console.log(offers);
     return offers.map(offer =>
@@ -83,185 +87,7 @@ const retrieveOffersFromDb = (offers) => {
 };
 
 
-let acceptanceID = null;
-let buyerID = null;
-let offerId = null;
 
-
-function confirmOfferAcceptance() {
-    $('.btn-accept').on('click', function (e) {
-        e.preventDefault();
-        offerId = $(this).data("id");
-        console.log(offerId);
-        $.get(`${OFFERS_URL}/${offerId}`).then(function (res) {
-            console.log(res);
-            populateAcceptedOfferDiv(res);
-        })
-        $("#btn-confirm").attr("data-id", offerId);
-    })///END OF CONFIRM FUNCTION
-
-    function populateAcceptedOfferDiv(res) {
-        const id = res.id;
-        const offeror = res.offeror.id;
-        const offerAmount = res.offerAmount;
-        const loanType = res.loanType;
-        const appraisalWaiver = res.appraisalWaiver;
-        const survey = res.survey;
-        const closingCosts = res.closingCosts;
-        const closingDate = res.closingDate;
-        const homeWarranty = res.homeWarranty;
-        const buyersAgent = res.offeror.buyerAgentID;
-        const acceptanceDate = new Date().toISOString().slice(0, 10)
-
-        acceptanceID = res.listing.id;
-        buyerID = res.offeror.id;
-
-        // //language=html
-        const acceptHTML = `
-				<div id="acceptOfferDiv" class="flex flex-wrap justify-evenly rounded m-1 bg-secondary">
-					<div class="text-center mx-1 my-2" id="offerId" data-id="${id}">
-						${id}
-					</div>
-					<div class="text-center mx-1 my-2" id="offerAmount-${id}">
-							\$${offerAmount}
-					</div>
-					<div id="closingCosts" class="text-center mx-1 my-2">
-						C/C: \$${closingCosts}
-					</div>
-					<div class="text-center mx-1 my-2">
-						Closing: ${closingDate}
-					</div>
-					<div class="text-center mx-1 my-2">
-						H/W: ${homeWarranty}
-					</div>
-					<div class="text-center mx-1 my-2">
-						L/T: ${loanType}
-					</div>
-					<div class="text-center mx-1 my-2">
-						Waive Appraisal: ${appraisalWaiver}
-					</div>
-					<div class="text-center mx-1 my-2">
-						Buyer Pays for Survey: ${survey}
-					</div>
-				</div>`
-
-        $("#offer").html("").append(`${acceptHTML}`);
-        $("#btn-confirm").removeClass("hidden");
-
-    }
-}/// END OF POPULATED ACCEPTED OFFER
-
-
-function initCounterOffer() {
-    $('.btn-counter').on('click', function () {
-        const id = $(this).data("id");
-        console.log(id);
-
-        $.get(`${OFFERS_URL}/${id}`).then(function (res) {
-            console.log(res);
-            populateCounterOfferForm(res);
-        })
-    })
-}///END OF COUNTER OFFER FUNCTION
-
-function populateCounterOfferForm(res) {
-    const id = res.id;
-    const offeror = res.offeror.id;
-    const offerAmount = res.offerAmount;
-    const loanType = res.loanType;
-    const appraisalWaiver = res.appraisalWaiver;
-    const survey = res.survey;
-    const closingCosts = res.closingCosts;
-    const closingDate = res.closingDate;
-    const homeWarranty = res.homeWarranty;
-    const buyersAgent = res.offeror.buyerAgentID;
-
-    //language=html
-    const acceptHTML = `
-		<div id="acceptOfferDiv" class="flex flex-wrap justify-evenly border-2 rounded border-black m-1">
-			<div class="text-center mx-1 my-2" id="offerId" data-id="${id}">
-				<label for="offerId">Offer ID</label>
-				<input type="text" readonly class="form-control" id="offerId" placeholder="${id}">
-			</div>
-			<div class="text-center mx-1 my-2">
-				<label for="offerLoanType">Loan Type:</label>
-				<input type="text" readonly class="form-control" id="offerLoanType" placeholder="${loanType}">
-			</div>
-			<div class="text-center mx-1 my-2">
-				<label for="offerAppWaiver">Waive Appraisal</label>
-				<input type="text" readonly class="form-control" id="offerAppWaiver" placeholder="${appraisalWaiver}">
-			</div>
-			<div class="text-center mx-1 my-2" id="offerAmount-${id}">
-				<label for="offerAmount">Offer Amt</label>
-				<input type="text" class="form-control" id="offerAmount" placeholder="${offerAmount}">
-			</div>
-			<div id="closingCosts" class="text-center mx-1 my-2">
-				<label for="offerClosingCosts">Closing Costs</label>
-				<input type="text" class="form-control" id="offerClosingCosts" placeholder="${closingCosts}">
-			</div>
-			<div class="text-center mx-1 my-2">
-				<label for="offerClosingDate">Closing Date</label>
-				<input type="text" class="form-control" id="offerClosingDate" placeholder="${closingDate}">
-			</div>
-			<div class="text-center mx-1 my-2">
-				<label for="offerHomeWarranty">Home Warranty</label>
-				<input type="text" class="form-control" id="offerHomeWarranty" placeholder="${homeWarranty}">
-			</div>
-			<div class="text-center mx-1 my-2">
-				<label for="offerSurvey">Survey</label>
-				<input type="text" class="form-control" id="offerSurvey" placeholder="${survey}">
-			</div>
-		</div>`
-
-    $("#offer").html("").append(`${acceptHTML}`);
-    $("#btn-confirm").removeClass("hidden");
-
-} /// END OF POPULATE CO FORM
-
-function updateOfferStatus() {
-///conditional logic is flawed here.  will not iterate
-    $("#btn-confirm").on('click', function () {
-        offerId = $("#btn-confirm").data("id");
-        const offerUpdate = {
-            method: "PUT",
-            headers: {
-                "Content-Type": 'application/json'
-            },
-        }
-        for (let i = 0; i < idArray.length; i++) {
-            let indexToDecline = idArray[i];
-            if (idArray[i] === offerId) {
-                fetch(`${OFFERS_URL}/${offerId}`, offerUpdate).then(function (res) {
-                    console.log(res)
-                })
-            } else {
-                fetch(`${OFFERS_URL}/decline/${indexToDecline}`, offerUpdate).then(function (res) {
-                    console.log(res)
-                })
-            }
-
-        }
-    })
-}
-
-function updateListingObject() {
-    $("#btn-confirm").on('click', function (e) {
-        e.preventDefault();
-        const soldListing = {
-            buyerID: buyerID,
-            // buyerAgentId: `${buyer.buyerAgentId}`
-        }
-        const listingUpdate = {
-            method: "PUT",
-            headers: {
-                "Content-Type": 'application/json'
-            },
-            body: JSON.stringify(soldListing)
-        }
-        fetch(`http://${BACKEND_HOST}:${PORT}/api/listings/acceptOffer/${acceptanceID}`, listingUpdate)
-            .then(response => createView(`/listing/listings/${acceptanceID}`))
-    })
-}////END OF UPDATE LISTING OBJECT
 
 
 // // RENDERING BUTTONS BASED ON USER IDENTITY && ACTIVITY
@@ -302,7 +128,7 @@ export function OfferEvent() {
     // areThereAnyListings();
     // renderMakeOfferBtn(props);
     confirmOfferAcceptance();
-    updateOfferStatus();
+    updateOfferStatus(idArray);
     updateListingObject();
     createMakeOfferView();
     initCounterOffer();
